--- ProcessLib/MassTransportProcess-fwd.h
@@ -1,18 +1,0 @@
-/**
- * \copyright
- * Copyright (c) 2012-2016, OpenGeoSys Community (http://www.opengeosys.org)
- *            Distributed under a Modified BSD License.
- *              See accompanying file LICENSE.txt or
- *              http://www.opengeosys.org/project/license
- *
- */
-
-#ifndef PROCESS_LIB_MASSTRANSPORTPROCESS_FWD_H_
-#define PROCESS_LIB_MASSTRANSPORTPROCESS_FWD_H_
-
-#include "MassTransportProcess.h"
-#include "NumericsConfig.h"
-
-extern template class ProcessLib::MassTransportProcess<GlobalSetupType>;
-
-#endif  // PROCESS_LIB_GROUNDWATERFLOWPROCESS_FWD_H_
