--- ProcessLib/MassTransportProcess.cpp
@@ -1,18 +1,0 @@
-/**
- * \copyright
- * Copyright (c) 2012-2016, OpenGeoSys Community (http://www.opengeosys.org)
- *            Distributed under a Modified BSD License.
- *              See accompanying file LICENSE.txt or
- *              http://www.opengeosys.org/project/license
- *
- */
-
-#include "MassTransportProcess-fwd.h"
-#include "MassTransportProcess.h"
-
-namespace ProcessLib
-{
-
-template class MassTransportProcess<GlobalSetupType>;
-
-}   // namespace ProcessLib
